# perf: Doubling and mixed doubling in Jacobians and ext. Jacobians