--- a/ecc/bls12-381/g1.go
+++ b/ecc/bls12-381/g1.go
@@ -78,6 +78,7 @@ func (p *G1Affine) ScalarMultiplicationBase(s *big.Int) *G1Affine {
 // It uses the Jacobian addition with a.Z=b.Z=1 and converts the result to affine coordinates.
 //
 // https://www.hyperelliptic.org/EFD/g1p/auto-shortw-jacobian-0.html#addition-mmadd-2007-bl
+// ~Cost: 4M + 2S
 func (p *G1Affine) Add(a, b *G1Affine) *G1Affine {
 	var q G1Jac
 	// a is infinity, return b
@@ -126,6 +127,7 @@ func (p *G1Affine) Add(a, b *G1Affine) *G1Affine {
 // addition with a.Z=1, and converts it back to affine coordinates.
 //
 // http://www.hyperelliptic.org/EFD/g1p/auto-shortw-jacobian-0.html#doubling-mdbl-2007-bl
+// ~Cost: 1M + 5S
 func (p *G1Affine) Double(a *G1Affine) *G1Affine {
 	var q G1Jac
 	q.FromAffine(a)
@@ -261,7 +263,8 @@ func (p *G1Jac) Neg(q *G1Jac) *G1Jac {
 
 // AddAssign sets p to p+a in Jacobian coordinates.
 //
-// https://hyperelliptic.org/EFD/g1p/auto-shortw-jacobian-3.html#addition-add-2007-bl
+// https://hyperelliptic.org/EFD/g1p/auto-shortw-jacobian-0.html#addition-add-2007-bl
+// ~Cost: 11M + 5S
 func (p *G1Jac) AddAssign(q *G1Jac) *G1Jac {
 
 	// p is infinity, return q
@@ -325,7 +328,8 @@ func (p *G1Jac) SubAssign(q *G1Jac) *G1Jac {
 
 // Double sets p to [2]q in Jacobian coordinates.
 //
-// https://hyperelliptic.org/EFD/g1p/auto-shortw-jacobian-3.html#doubling-dbl-2007-bl
+// https://www.hyperelliptic.org/EFD/g1p/auto-shortw-jacobian-0.html#doubling-mdbl-2007-bl
+// ~Cost: 1M + 5S
 func (p *G1Jac) DoubleMixed(a *G1Affine) *G1Jac {
 	var XX, YY, YYYY, S, M, T fp.Element
 	XX.Square(&a.X)
@@ -356,6 +360,7 @@ func (p *G1Jac) DoubleMixed(a *G1Affine) *G1Jac {
 // AddMixed sets p to p+a in Jacobian coordinates, where a.Z = 1.
 //
 // http://www.hyperelliptic.org/EFD/g1p/auto-shortw-jacobian-0.html#addition-madd-2007-bl
+// ~Cost: 7M + 4S
 func (p *G1Jac) AddMixed(a *G1Affine) *G1Jac {
 
 	//if a is infinity return p
@@ -405,7 +410,8 @@ func (p *G1Jac) AddMixed(a *G1Affine) *G1Jac {
 
 // Double sets p to [2]q in Jacobian coordinates.
 //
-// https://hyperelliptic.org/EFD/g1p/auto-shortw-jacobian-3.html#doubling-dbl-2007-bl
+// https://www.hyperelliptic.org/EFD/g1p/auto-shortw-jacobian-0.html#doubling-dbl-2009-l
+// ~Cost: 2M + 5S
 func (p *G1Jac) Double(q *G1Jac) *G1Jac {
 	p.Set(q)
 	p.DoubleAssign()
@@ -414,33 +420,31 @@ func (p *G1Jac) Double(q *G1Jac) *G1Jac {
 
 // DoubleAssign doubles p in Jacobian coordinates.
 //
-// https://hyperelliptic.org/EFD/g1p/auto-shortw-jacobian-3.html#doubling-dbl-2007-bl
+// https://www.hyperelliptic.org/EFD/g1p/auto-shortw-jacobian-0.html#doubling-dbl-2009-l
+// ~Cost: 2M + 5S
 func (p *G1Jac) DoubleAssign() *G1Jac {
-
-	var XX, YY, YYYY, ZZ, S, M, T fp.Element
-
-	XX.Square(&p.X)
-	YY.Square(&p.Y)
-	YYYY.Square(&YY)
-	ZZ.Square(&p.Z)
-	S.Add(&p.X, &YY)
-	S.Square(&S).
-		Sub(&S, &XX).
-		Sub(&S, &YYYY).
-		Double(&S)
-	M.Double(&XX).Add(&M, &XX)
-	p.Z.Add(&p.Z, &p.Y).
-		Square(&p.Z).
-		Sub(&p.Z, &YY).
-		Sub(&p.Z, &ZZ)
-	T.Square(&M)
-	p.X = T
-	T.Double(&S)
-	p.X.Sub(&p.X, &T)
-	p.Y.Sub(&S, &p.X).
-		Mul(&p.Y, &M)
-	YYYY.Double(&YYYY).Double(&YYYY).Double(&YYYY)
-	p.Y.Sub(&p.Y, &YYYY)
+	var A, B, C, D, E, F, t fp.Element
+	A.Square(&p.X)
+	B.Square(&p.Y)
+	C.Square(&B)
+	D.Add(&p.X, &B).
+		Square(&D).
+		Sub(&D, &A).
+		Sub(&D, &C).
+		Double(&D)
+	E.Double(&A).
+		Add(&E, &A)
+	F.Square(&E)
+	t.Double(&D)
+	p.Z.Mul(&p.Y, &p.Z).
+		Double(&p.Z)
+	p.X.Sub(&F, &t)
+	p.Y.Sub(&D, &p.X).
+		Mul(&p.Y, &E)
+	t.Double(&C).
+		Double(&t).
+		Double(&t)
+	p.Y.Sub(&p.Y, &t)
 
 	return p
 }
@@ -839,6 +843,7 @@ func (p *G1Jac) unsafeFromJacExtended(q *g1JacExtended) *G1Jac {
 // add sets p to p+q in extended Jacobian coordinates.
 //
 // https://www.hyperelliptic.org/EFD/g1p/auto-shortw-xyzz.html#addition-add-2008-s
+// ~Cost: 12M + 2S
 func (p *g1JacExtended) add(q *g1JacExtended) *g1JacExtended {
 	//if q is infinity return p
 	if q.ZZ.IsZero() {
@@ -896,6 +901,8 @@ func (p *g1JacExtended) add(q *g1JacExtended) *g1JacExtended {
 // double sets p to [2]q in Jacobian extended coordinates.
 //
 // http://www.hyperelliptic.org/EFD/g1p/auto-shortw-xyzz.html#doubling-dbl-2008-s-1
+// ~Cost: 6M + 3S
+//
 // N.B.: since we consider any point on Z=0 as the point at infinity
 // this doubling formula works for infinity points as well.
 func (p *g1JacExtended) double(q *g1JacExtended) *g1JacExtended {
@@ -925,6 +932,7 @@ func (p *g1JacExtended) double(q *g1JacExtended) *g1JacExtended {
 // addMixed sets p to p+q in extended Jacobian coordinates, where a.ZZ=1.
 //
 // http://www.hyperelliptic.org/EFD/g1p/auto-shortw-xyzz.html#addition-madd-2008-s
+// ~Cost: 8M + 2S
 func (p *g1JacExtended) addMixed(a *G1Affine) *g1JacExtended {
 
 	//if a is infinity return p
@@ -981,6 +989,7 @@ func (p *g1JacExtended) addMixed(a *G1Affine) *g1JacExtended {
 // subMixed works the same as addMixed, but negates a.Y.
 //
 // http://www.hyperelliptic.org/EFD/g1p/auto-shortw-xyzz.html#addition-madd-2008-s
+// ~Cost: 8M + 2S
 func (p *g1JacExtended) subMixed(a *G1Affine) *g1JacExtended {
 
 	//if a is infinity return p
@@ -1035,27 +1044,29 @@ func (p *g1JacExtended) subMixed(a *G1Affine) *g1JacExtended {
 
 }
 
-// doubleNegMixed works the same as double, but negates q.Y.
+// doubleNegMixed works the same as doubleMixed, but negates q.Y.
+//
+// https://www.hyperelliptic.org/EFD/g1p/auto-shortw-xyzz.html#doubling-mdbl-2008-s-1
+// ~Cost: 4M + 3S
 func (p *g1JacExtended) doubleNegMixed(a *G1Affine) *g1JacExtended {
 
-	var U, V, W, S, XX, M, S2, L fp.Element
+	var U, V, W, S, M, t fp.Element
 
 	U.Double(&a.Y)
 	U.Neg(&U)
 	V.Square(&U)
 	W.Mul(&U, &V)
 	S.Mul(&a.X, &V)
-	XX.Square(&a.X)
-	M.Double(&XX).
-		Add(&M, &XX) // -> + A, but A=0 here
-	S2.Double(&S)
-	L.Mul(&W, &a.Y)
-
-	p.X.Square(&M).
-		Sub(&p.X, &S2)
+	t.Square(&a.X)
+	M.Double(&t).
+		Add(&M, &t) // -> + A, but A=0 here
+	p.X.Square(&M)
+	t.Double(&S)
+	p.X.Sub(&p.X, &t)
+	t.Mul(&W, &a.Y)
 	p.Y.Sub(&S, &p.X).
 		Mul(&p.Y, &M).
-		Add(&p.Y, &L)
+		Add(&p.Y, &t)
 	p.ZZ.Set(&V)
 	p.ZZZ.Set(&W)
 
@@ -1064,26 +1075,26 @@ func (p *g1JacExtended) doubleNegMixed(a *G1Affine) *g1JacExtended {
 
 // doubleMixed sets p to [2]a in Jacobian extended coordinates, where a.ZZ=1.
 //
-// http://www.hyperelliptic.org/EFD/g1p/auto-shortw-xyzz.html#doubling-dbl-2008-s-1
+// https://www.hyperelliptic.org/EFD/g1p/auto-shortw-xyzz.html#doubling-mdbl-2008-s-1
+// ~Cost: 4M + 3S
 func (p *g1JacExtended) doubleMixed(a *G1Affine) *g1JacExtended {
 
-	var U, V, W, S, XX, M, S2, L fp.Element
+	var U, V, W, S, M, t fp.Element
 
 	U.Double(&a.Y)
 	V.Square(&U)
 	W.Mul(&U, &V)
 	S.Mul(&a.X, &V)
-	XX.Square(&a.X)
-	M.Double(&XX).
-		Add(&M, &XX) // -> + A, but A=0 here
-	S2.Double(&S)
-	L.Mul(&W, &a.Y)
-
-	p.X.Square(&M).
-		Sub(&p.X, &S2)
+	t.Square(&a.X)
+	M.Double(&t).
+		Add(&M, &t) // -> + A, but A=0 here
+	p.X.Square(&M)
+	t.Double(&S)
+	p.X.Sub(&p.X, &t)
+	t.Mul(&W, &a.Y)
 	p.Y.Sub(&S, &p.X).
 		Mul(&p.Y, &M).
-		Sub(&p.Y, &L)
+		Sub(&p.Y, &t)
 	p.ZZ.Set(&V)
 	p.ZZZ.Set(&W)