… ecc/bls24-315/g1.go
… ecc/bls24-315/g2.go
… ecc/bls24-317/g1.go
… ecc/bls24-317/g2.go
… ecc/bn254/g1.go
… ecc/bn254/g2.go
… ecc/bw6-633/g1.go
… ecc/bw6-633/g2.go
… ecc/bw6-761/g1.go
… ecc/bw6-761/g2.go
… ecc/grumpkin/g1.go
… ecc/secp256k1/g1.go
… internal/generator/ecc/template/point.go.tmpl