# Description

This PR corrects comments (URL to EFD forumulas) and saves `2m` in Jacobian doubling and `3m` in extended Jacobian mixed doubling.


## Type of change

- [x] New feature (non-breaking change which adds functionality)

# How has this been tested?

Current tests pass.

# How has this been benchmarked?

e.g. BLS12-381 on MBA M1:
```
benchmark                  old ns/op     new ns/op     delta
BenchmarkG1JacDouble-8     338           298           -11.76%
```


# Checklist:

- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I did not modify files generated from templates
- [x] `golangci-lint` does not output errors locally
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream modules